fix profile view, tests and hasRated box primitive

+++ persistence/src/test/java/ar/edu/itba/paw/persistence/hibernate/CommentHibernateDaoTest.java
@@ -83,8 +83,8 @@ public class CommentHibernateDaoTest {
         Assert.assertNotNull(ans);
         Assert.assertEquals(1,ans.size());
         Assert.assertEquals(ID+6,ans.get(0).getId().longValue());
-        Assert.assertEquals(ID.longValue(),ans.get(0).getReceiver().getId());
-        Assert.assertEquals(ID+1,ans.get(0).getCommenter().getId());
+        Assert.assertEquals(ID.longValue(),ans.get(0).getReceiver().getId().longValue());
+        Assert.assertEquals(ID+1,ans.get(0).getCommenter().getId().longValue());
         Assert.assertEquals("First",ans.get(0).getComment());
         Assert.assertEquals(LocalDateTime.of(2025,11,16,22,4,7),ans.get(0).getCreatedAt());
     }
@@ -97,8 +97,8 @@ public class CommentHibernateDaoTest {
         Assert.assertEquals(COMMENTS_PAGE_SIZE,ans.size());
         for (int i = 0; i < COMMENTS_PAGE_SIZE; i++) {
             Assert.assertEquals(ID+i,ans.get(i).getId().longValue());
-            Assert.assertEquals(ID.longValue(),ans.get(i).getReceiver().getId());
-            Assert.assertEquals(ID+1,ans.get(i).getCommenter().getId());
+            Assert.assertEquals(ID.longValue(),ans.get(i).getReceiver().getId().longValue());
+            Assert.assertEquals(ID+1,ans.get(i).getCommenter().getId().longValue());
             Assert.assertEquals("First",ans.get(i).getComment());
             Assert.assertEquals(LocalDateTime.of(2025,11,16,22,4,1+i),ans.get(i).getCreatedAt());
         }
